requirements.txt
# Melanoma Detection with Hybrid GWO-ANN
# Python >= 3.8

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing & Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image Processing
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bar
tqdm>=4.65.0

# Optional: For GPU support
# Install PyTorch with CUDA from: https://pytorch.org/get-started/locally/
# Example: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

